// src/propuestas/dto/actualizar-propuesta.dto.ts
export class UpdatePropuestaDto {    
    id_Propuesta?: number;
    id_Acreditado?: number;
    f_RecepcionINDEP?: string;
    f_RecepcionCRCA?: string;
    f_RecepcionDocumento?: string;
    id_TipoPropuesta?: number;
    d_MontoMN?: number;
    d_MontoUSD?: number;
    s_MontoLetraMN?: string;
    s_MontoLetraUSD?: string;
    d_ValorCOPER?: string;
    i_PlazoTotal?: number;
    s_NumeroVolante?: number;
    s_Observaciones?: string;
    id_Usuario?: number;
    id_Especialista?: number;
    f_Captura?: string;
    swc_Activo?:boolean;  
}